Reject out-of-range token IDs instead of trapping
Config.integer() returns full-range Int, and the init used trapping
Int32(_:) conversions on IDs read from tokenizer.json - so a corrupt or
foreign tokenizer directory (TURBO_FIELDFARE_TOKENIZER_DIR can point
anywhere) with an added-token ID outside Int32 range crashed the process
at load instead of taking the typed-error path every other
malformed-config case uses. requireTokenID now returns a checked Int32
and the specials/bos/eos conversions go through the same guard, throwing
GFTokenizerError.invalidTokenID.

Also merges the two identical test ID helpers into one tokenID(_:).

+++ Tests/TurboFieldfareServer/OpenAIValidationTests.swift
@@ -514,7 +514,7 @@ struct GemmaToolCallTests {
         try feed(tokenizer.channelStartID)
         for id in tokenizer.encode("thought\n", addBOS: false) { try feed(id) }
         for token in ["<0xF0>", "<0x9F>", "<0x98>", "<0x80>"] {
-            try feed(Int32(GFTokenizer.requireTokenID(tokenizer.tokenizer, token)))
+            try feed(GFTokenizer.requireTokenID(tokenizer.tokenizer, token))
         }
         try feed(tokenizer.channelEndID)
         for id in tokenizer.encode("ok", addBOS: false) { try feed(id) }
